Reprint game over itself, track user guesses, and filter input.
Using ANSI escape codes we can move our cursor up/down the stdout,
allowing us to reprint over our previously printed lines.
User's previous guesses are tracked in a list and printed to screen.
User input is now filtered by length > 1 and regex is used to determine
if the input is composed of characters in the alphabet.